Move Resources into own component
